+++ src/demo/kotlin/com/loadingbyte/cinecred/demos/ScreencastDemos.kt
@@ -115,7 +115,7 @@ object ScreencastScreencastDemo : ScreencastDemo(
         val creditsFile = projectDir.resolve("Credits.csv")
         val spreadsheetEditorWin = SpreadsheetEditorVirtualWindow(creditsFile, skipRows = 1).apply {
             size = Dimension(800, 500)
-            colWidths = intArrayOf(160, 160, 50, 110, 110, 80, 80, 80, 110)
+            colWidths = intArrayOf(160, 160, 50, 110, 110, 80, 80, 80, 80, 110)
         }
         dt.add(spreadsheetEditorWin)
         dt.center(spreadsheetEditorWin)